requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
google-generativeai>=0.8.0
